Kinvey Connection Rerouted to Main Activity
Only the Main Activity can establish client connection to kinvey but
the any activity can use that single Client to communicate with Kinvey.

Added ActivityConstants interface to contain constant integer ID values
for each Activity.

Added a KinveyConnectionSingleton class that will be a single instance
of the client connection to Kinvey.