Fixed most redundancy, all is centralized now
Kinvey communication and data download are now centralized in the
KinveyConnectionTask and results are assigned to proper classes through
the KinveyCaller.
Consequently, there are only single, static instances of kinvey
querying methods that all classes can use.

Updated the stock class to the final list of ratios.
It still contains some redundant data that wasn’t removed to prevent
database conflicts.
All ratios are now floats so the StockDataSource was updated
correspondingly.

Fixed the hot stocks fragment to download and show the correct grades.

Fixed each grade segment (Dividend, Growth, Total) to display correct
data.

All constants, names, and other values that are standardized throughout
the application are in the KinveyConstants class.

The autocomplete now works only with tickers but it has an underlying
Trie data structure that makes autocomplete practically instantaneous.