Updated KinveyConstants
KinveyConstants were updated to ApplicationConstants and moved to the
main activity fragment since it made more sense to have all constants
relative to the application at one place.

Next step is to update the DisplayStockRatioData class to fix text
spacing and alignment and add a third row of sector information to the
ratio table.
This sector information will be used to identify how does each
particular ratio of a company compare to the average value for that
ratio in the whole sector.
Also coloring needs to be added when a ratio is better than average
(Green) or worse (Red).