Added Portfolio fragment
Added a new fragment that will contain the user’s stock Portfolio.
It is at second position while the market fragment is moved to the last
position and we are considering if we should keep it since this
information isn’t very unique and is provided by most other apps and
many other sources.