Portfolio fragment is fully functional
Finished the portfolio fragment and added the addToPortfolio
functionality on the ratio display page.

The portfolio is saved in sharedPreferences so it stays remembered on
the phone even if the application gets killed.

The stocks are displayed in sorted order by ticker.

Also added on-click-highlighting for all links in the app that can be
clicked.