small change when scores are duplicates
Added a small change in DisplayStockRatioData class that append the
first character of the score (for example ’t’ for total score) to the
actual score to be able to distinguish equivalent scores since Set<> in
java doesn’t allow duplicates.

Consequently, on the portfolio page, we display a substring excluding
that extra character.