Aligned column names on the portfolio page
removed the hard-coded column names from the portfolio fragment XML
file and added method addColumnNames() in the portfolio fragment code
that inserts the column names programmatically.
The reason for this was to align the columns properly since before they
were in a separate table above our portfolio table.
Now they are getting redrawn with the whole table on each change.